requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements/requirements-base.in
#
--trusted-host pypi.python.org
--trusted-host pypi.org
--trusted-host files.pythonhosted.org

attrs==20.3.0
    # via -r requirements/requirements-base.in
cx-oracle==8.1.0
    # via -r requirements/requirements-base.in
et-xmlfile==1.0.1
    # via openpyxl
jdcal==1.4.1
    # via openpyxl
numpy==1.20.1
    # via
    #   pandas
    #   scipy
openpyxl==3.0.6
    # via -r requirements/requirements-base.in
pandas==1.1.5
    # via -r requirements/requirements-base.in
python-dateutil==2.8.1
    # via pandas
python-dotenv==0.15.0
    # via -r requirements/requirements-base.in
pytz==2021.1
    # via pandas
scipy==1.6.0
    # via -r requirements/requirements-base.in
six==1.15.0
    # via python-dateutil
sqlalchemy==1.3.22
    # via -r requirements/requirements-base.in
